refactor(enter-text): remove legacy input paths
