+++ benchmark/adbandroid/bridge/tools_api.py
@@ -209,32 +209,15 @@ class ADBToolsAPIHandler:
                 },
             },
             {
-                "name": "enter_text_in_field",
-                "description": "Enter target text into an input field on the Android device: taps the focus point, then types via adb input text. Accepts the Go agent text-entry contract.",
+                "name": "enter_text",
+                "description": "Enter target text into an input field on the Android device: taps the focus point, then types via adb input text.",
                 "args_schema": {
                     "type": "object",
                     "additionalProperties": False,
                     "properties": {
                         "text": {"type": "string", "description": "Exact text that must appear in the field."},
-                        "platform": {"type": "string", "enum": ["ios", "android", "mac"]},
-                        "mode": {"type": "string", "enum": ["form", "search"]},
                         "focus": enter_text_focus_schema,
-                        "segments": {"type": "array", "items": {"type": "string"}},
-                        "max_attempts": {"type": "integer", "minimum": 1},
-                    },
-                    "required": ["text", "focus"],
-                },
-            },
-            {
-                "name": "enter_text_via_bridge",
-                "description": "ADB-compatible alias for text entry. Types via adb input text instead of the physical phone bridge or HID clipboard path.",
-                "args_schema": {
-                    "type": "object",
-                    "additionalProperties": False,
-                    "properties": {
-                        "text": {"type": "string", "description": "Exact text that must appear in the field."},
-                        "platform": {"type": "string", "enum": ["ios", "android", "mac"]},
-                        "focus": enter_text_focus_schema,
+                        "send_after_commit": {"type": "boolean"},
                     },
                     "required": ["text", "focus"],
                 },
@@ -333,7 +316,7 @@ class ADBToolsAPIHandler:
                 tool_input = {}
             else:
                 parsed_input = json.loads(raw_input)
-                if tool_name in ("keyboard_text", "enter_text_in_field", "enter_text_via_bridge") and isinstance(parsed_input, str):
+                if tool_name == "keyboard_text" and isinstance(parsed_input, str):
                     tool_input = {"text": parsed_input}
                 elif isinstance(parsed_input, dict):
                     tool_input = parsed_input
@@ -345,7 +328,7 @@ class ADBToolsAPIHandler:
                     )
                     return
         except json.JSONDecodeError:
-            if tool_name in ("keyboard_text", "enter_text_in_field", "enter_text_via_bridge"):
+            if tool_name == "keyboard_text":
                 tool_input = {"text": raw_input.strip()}
             else:
                 self._send_json(
@@ -436,10 +419,8 @@ class ADBToolsAPIHandler:
             return self._call_keyboard_text(tool_input)
         elif tool_name == "keyboard_tap":
             return self._call_keyboard_tap(tool_input)
-        elif tool_name == "enter_text_in_field":
-            return self._call_enter_text(tool_input, tool_name="enter_text_in_field")
-        elif tool_name == "enter_text_via_bridge":
-            return self._call_enter_text(tool_input, tool_name="enter_text_via_bridge")
+        elif tool_name == "enter_text":
+            return self._call_enter_text(tool_input)
         elif tool_name == "mouse_click":
             return self._call_mouse_click(tool_input)
         elif tool_name == "mouse_move":
@@ -707,39 +688,23 @@ class ADBToolsAPIHandler:
 
         return {"output": f"error: unsupported quick_action: {tool_input.get('action')!r}", "is_error": True}
 
-    def _call_enter_text(self, tool_input: dict[str, Any], *, tool_name: str) -> dict[str, Any]:
-        """Execute Go-agent text-entry tools through adb input text."""
+    def _call_enter_text(self, tool_input: dict[str, Any]) -> dict[str, Any]:
+        """Execute enter_text through adb input text."""
         text = tool_input.get("text", "")
         if not isinstance(text, str):
             return {"output": "error: text must be a string", "is_error": True}
         if text == "":
-            output = {
-                "ok": False,
-                "committed": False,
-                "target_text": "",
-                "required_mode": "ascii",
-                "attempts": 1,
-                "ime_switches": 0,
-                "vlm_calls": 0,
-                "reason": "text is required",
-            }
+            output = {"ok": False, "suggestion": "Provide non-empty text, then retry enter_text."}
             return {"output": json.dumps(output), "is_error": False}
         unsupported = unsupported_adb_text_chars(text)
         if unsupported:
             output = {
                 "ok": False,
-                "committed": False,
-                "target_text": text,
-                "required_mode": _required_text_input_mode(text),
-                "attempts": 1,
-                "ime_switches": 0,
-                "vlm_calls": 0,
-                "reason": f"adb input text supports only US-keyboard ASCII characters; unsupported: {unsupported!r}",
+                "suggestion": f"adb input text supports only US-keyboard ASCII characters; unsupported: {unsupported!r}",
             }
             return {"output": json.dumps(output, ensure_ascii=False), "is_error": False}
 
         device = self.state.device
-        steps: list[str] = []
         focus = tool_input.get("focus")
         focus_pixels: tuple[int, int] | None = None
         if focus is not None:
@@ -760,34 +725,19 @@ class ADBToolsAPIHandler:
         def enter_text() -> None:
             if focus_pixels is not None:
                 device.tap(*focus_pixels)
-                steps.append(f"tap focus {focus_pixels[0]} {focus_pixels[1]}")
                 time.sleep(FOCUS_SETTLE_SEC)
             device.input_text(text)
-            steps.append("adb input text")
 
         result = self._execute_device(
             enter_text,
-            tool_name=tool_name,
+            tool_name="enter_text",
             tool_input=tool_input,
             adb_summary="input text",
         )
         if result.get("is_error"):
             return result
 
-        mode = _text_input_mode(tool_input.get("mode"))
-        output = {
-            "ok": True,
-            "committed": True,
-            "target_text": text,
-            "field_text": text,
-            "required_mode": _required_text_input_mode(text),
-            "mode": mode,
-            "attempts": 1,
-            "ime_switches": 0,
-            "vlm_calls": 0,
-            "reason": "adb input text sent",
-            "steps": steps,
-        }
+        output = {"ok": True}
         return {"output": json.dumps(output, ensure_ascii=False), "is_error": False}
 
     def _call_keyboard_text(self, tool_input: dict[str, Any]) -> dict[str, Any]:
@@ -1132,19 +1082,6 @@ def _finite_float(value: Any, name: str) -> float:
 
 def _clamp(value: float, minimum: float, maximum: float) -> float:
     return max(minimum, min(maximum, value))
-
-
-def _required_text_input_mode(text: str) -> str:
-    if any(ord(ch) > 127 for ch in text):
-        return "composition"
-    return "ascii"
-
-
-def _text_input_mode(value: Any) -> str:
-    mode = str(value or "").strip().lower()
-    if mode == "search":
-        return "search"
-    return "form"
 
 
 def _quick_action_id(tool_input: dict[str, Any]) -> str:
