+++ benchmark/docker/docker-compose.agent-daemon.yml
@@ -15,7 +15,7 @@ services:
       - ${AIDEN_CONFIG_DIR:-../config}:/config:ro
     environment:
       AIDEN_ENVIRONMENT_BRIDGE_MODE: ${AIDEN_ENVIRONMENT_BRIDGE_MODE:-0}
-      AIDEN_ENVIRONMENT_BRIDGE_TOOLS: "screenshot,touch_gesture,keyboard_text,keyboard_tap,enter_text_in_field,enter_text_via_bridge,mouse_click,mouse_move,mouse_scroll,quick_action"
+      AIDEN_ENVIRONMENT_BRIDGE_TOOLS: "screenshot,touch_gesture,keyboard_text,keyboard_tap,enter_text,mouse_click,mouse_move,mouse_scroll,quick_action"
       ENVIRONMENT_BRIDGE_ENDPOINT: ${ENVIRONMENT_BRIDGE_ENDPOINT:-}
       AIDEN_BENCHMARK_TASK_ID: ${AIDEN_BENCHMARK_TASK_ID:-}
       AIDEN_BENCHMARK_TOKEN_FILE: ${AIDEN_BENCHMARK_TOKEN_FILE:-}
