+++ benchmark/suites/phone_control_v1.json
@@ -9,7 +9,7 @@
       "skill_name": "device-operator"
     }
   ],
-  "prompt_prefix": "你正在操作一台通过 HDMI 采集画面的手机，不是 macOS、Linux 或桌面系统；测试环境可能是 iOS、Android 或其他移动系统。开始操作目标 UI 前，必须先调用 skill_read 读取 device-operator；读取后优先使用 quick_action 等语义工具，再根据需要使用 screenshot、touch_gesture、mouse_click、enter_text_in_field、keyboard_text、keyboard_tap 等手机 UI 工具。不要使用 shell、open、osascript 或桌面命令来打开目标手机上的 App。每次工具返回截图后，先判断页面是否变化、是否已经进入目标页面，再继续下一步；如果截图没有变化，不要反复点击同一坐标。打开系统设置或时钟时，优先点击当前手机主屏幕上可见的图标，或使用系统搜索。",
+  "prompt_prefix": "你正在操作一台通过 HDMI 采集画面的手机，不是 macOS、Linux 或桌面系统；测试环境可能是 iOS、Android 或其他移动系统。开始操作目标 UI 前，必须先调用 skill_read 读取 device-operator；读取后优先使用 quick_action 等语义工具，再根据需要使用 screenshot、touch_gesture、mouse_click、enter_text、keyboard_text、keyboard_tap 等手机 UI 工具。不要使用 shell、open、osascript 或桌面命令来打开目标手机上的 App。每次工具返回截图后，先判断页面是否变化、是否已经进入目标页面，再继续下一步；如果截图没有变化，不要反复点击同一坐标。打开系统设置或时钟时，优先点击当前手机主屏幕上可见的图标，或使用系统搜索。",
   "global_reset": {},
   "tasks": [
     {
