+++ benchmark/tests/test_adbandroid_bridge.py
@@ -236,8 +236,7 @@ def test_tools_catalog_lists_expected_tools(bridge):
         "touch_gesture",
         "keyboard_text",
         "keyboard_tap",
-        "enter_text_in_field",
-        "enter_text_via_bridge",
+        "enter_text",
         "mouse_click",
         "mouse_move",
         "mouse_scroll",
