+++ src/agent/internal/agent/skill_merge_test.go
@@ -895,9 +895,7 @@ func TestBundledDeviceOperatorAllowedToolsCoverEmbeddedPlaybooks(t *testing.T) {
 		"mouse_move",
 		"mouse_scroll",
 		"keyboard_tap",
-		"keyboard_text",
-		"enter_text_in_field",
-		"enter_text_via_bridge",
+		"enter_text",
 		"search_launch_app",
 		"request_human_handoff",
 		"recall_memory",
