--- src/agent/internal/agent/tools_enter_text_in_field.go
@@ -1,163 +1,0 @@
-package agent
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"strings"
-)
-
-type EnterTextInFieldTool struct {
-	engine               *textInputEngine
-	bridgeTool           *EnterTextViaBridgeTool
-	iosKeyboardIsolation *iosKeyboardIsolationController
-}
-
-func (t *EnterTextInFieldTool) Name() string { return "enter_text_in_field" }
-
-func (t *EnterTextInFieldTool) Description() string {
-	return `Enter text into a focused input field with automatic strategy selection and verification. ` +
-		`Prefer this over keyboard_text for any input field entry; keyboard_text is ASCII-only and has no field verification. ` +
-		`Precondition: the latest screenshot must clearly show the actual editable field or composer, and focus coordinates must be inside that visible field. Opening an app, folder/list view, or screen that only shows a create/new button is not enough; first create or open the document/message and observe its editor. Never use a guessed blank-space coordinate. ` +
-		`Use this for search fields, contact lookup, short text when no target-preserving Phone Bridge clipboard path is available, and normal IME/pinyin entry. When runtime app_text_entry_strategy is target_preserving_bridge, do not use this for non-search CJK/non-ASCII, multiline, or final composer text; call enter_text_via_bridge directly. For a user request to send, reply, or message from an already-open composer, set send_after_commit=true. ` +
-		`Returns committed:true only when vision confirms the target text is fully committed in the input field. field_text is diagnostic visual transcription and is not used for code-point equality. ` +
-		`Focus coordinates use the same coord_space system as touch/click tools; prefer normalized coordinates from the latest screenshot.`
-}
-
-func (t *EnterTextInFieldTool) ArgsSchema() map[string]any {
-	return objectArgsSchema(map[string]any{
-		"text":              stringArgSchema("Exact text that must appear in the field when done."),
-		"mode":              stringEnumArgSchema("Interaction mode. Use \"search\" for quick handoff in search boxes; omit for normal form entry.", "form", "search"),
-		"focus":             focusPointArgSchema("Coordinates inside an actual editable field or composer that is clearly visible in the latest screenshot; do not use blank space, an app/folder/list page, or a create/new button as the field."),
-		"segments":          stringArrayArgSchema("IME romanization syllables for CJK (e.g. [\"ni\",\"hao\"] for 你好). Pass [] for pure ASCII text."),
-		"max_attempts":      integerArgSchema("Retry attempts on verify failure (default 3)."),
-		"send_after_commit": boolArgSchema("After the exact target text is verified in the field, press send/submit and verify the input cleared or changed. Set true when the user asked to send, reply, or message and the target chat/composer is already open."),
-	}, "text", "focus", "segments")
-}
-
-func (t *EnterTextInFieldTool) Call(ctx context.Context, input string) (string, error) {
-	var controller *iosKeyboardIsolationController
-	if t != nil {
-		controller = t.iosKeyboardIsolation
-	}
-	return withIOSKeyboardIsolationBatchCall(ctx, controller, func(batchCtx context.Context) (string, error) {
-		return t.call(batchCtx, input)
-	})
-}
-
-func (t *EnterTextInFieldTool) call(ctx context.Context, input string) (string, error) {
-	if t == nil || t.engine == nil {
-		return toolErrorResultString(ctx, CodeModuleUnavailable, "enter_text_in_field is not fully configured"), nil
-	}
-	var args enterTextInFieldArgs
-	if err := json.Unmarshal([]byte(strings.TrimSpace(input)), &args); err != nil {
-		return toolErrorResultf(ctx, CodeInvalidArguments, "invalid input: %v", err), nil
-	}
-	if t.shouldPreferBridgeClipboard(args) {
-		bridgeResult, attempted := t.bridgeTool.runClipboardFirstResult(ctx, args)
-		if attempted && bridgeResult.Committed {
-			return jsonString(finalizeEnterTextInFieldResult(bridgeResult, args.SendAfterCommit)), nil
-		}
-		if attempted && strings.TrimSpace(bridgeResult.FieldText) != "" {
-			if strings.TrimSpace(bridgeResult.Reason) != "" {
-				bridgeResult.Reason += "; fresh observation required before corrective input"
-			} else {
-				bridgeResult.Reason = "fresh observation required before corrective input"
-			}
-			return jsonString(finalizeEnterTextInFieldResult(bridgeResult, args.SendAfterCommit)), nil
-		}
-		result, err := t.engine.Run(ctx, args)
-		if err != nil {
-			var toolErr *ToolError
-			if errors.As(err, &toolErr) {
-				SetToolError(ctx, toolErr)
-				return toolErrorString(toolErr), nil
-			}
-			return toolErrorResultf(ctx, CodeToolExecutionFailed, "%v", err), nil
-		}
-		if attempted {
-			result = mergeClipboardFallbackResult(bridgeResult, result)
-		}
-		return jsonString(finalizeEnterTextInFieldResult(result, args.SendAfterCommit)), nil
-	}
-	result, err := t.engine.Run(ctx, args)
-	if err != nil {
-		var toolErr *ToolError
-		if errors.As(err, &toolErr) {
-			SetToolError(ctx, toolErr)
-			return toolErrorString(toolErr), nil
-		}
-		return toolErrorResultf(ctx, CodeToolExecutionFailed, "%v", err), nil
-	}
-	return jsonString(finalizeEnterTextInFieldResult(result, args.SendAfterCommit)), nil
-}
-
-func finalizeEnterTextInFieldResult(result enterTextInFieldResult, sendAfterCommit bool) enterTextInFieldResult {
-	if !result.Committed {
-		result.OK = false
-		if result.Reason == "" {
-			result.Reason = "text entry not verified in field"
-		}
-		return result
-	}
-	if sendAfterCommit && !result.SendVerified {
-		result.OK = false
-		result.Reason = "field verified but send was not verified"
-		return result
-	}
-	result.OK = true
-	return result
-}
-
-func mergeClipboardFallbackResult(clipboardResult, fallbackResult enterTextInFieldResult) enterTextInFieldResult {
-	merged := fallbackResult
-	merged.VLMCalls += clipboardResult.VLMCalls
-	reason := strings.TrimSpace(clipboardResult.Reason)
-	if !merged.Committed && reason != "" {
-		fallbackReason := strings.TrimSpace(merged.Reason)
-		guidance := "clipboard path failed; continue with same target field and latest screenshot evidence"
-		if fallbackReason != "" {
-			merged.Reason = fallbackReason + "; safe clipboard failed earlier: " + reason + "; " + guidance
-		} else {
-			merged.Reason = "safe clipboard failed earlier: " + reason + "; " + guidance
-		}
-	}
-	return merged
-}
-
-func (t *EnterTextInFieldTool) shouldPreferBridgeClipboard(args enterTextInFieldArgs) bool {
-	if t == nil || t.bridgeTool == nil {
-		return false
-	}
-	var hw *textInputHardwareDeps
-	if t.engine != nil {
-		hw = &t.engine.hw
-	} else if t.bridgeTool != nil {
-		hw = t.bridgeTool.hw
-	}
-	platform := textInputHardwarePlatform(hw)
-	if platform != "ios" && platform != "android" {
-		return false
-	}
-	if strings.EqualFold(strings.TrimSpace(args.Mode), "search") {
-		return false
-	}
-	text := strings.TrimSpace(args.Text)
-	if text == "" || !textShouldPreferBridgeClipboard(text) {
-		return false
-	}
-	return t.bridgeTool.canUseClipboardFirst(platform, text)
-}
-
-func textShouldPreferBridgeClipboard(text string) bool {
-	if needsCompositionInput(text) {
-		return true
-	}
-	if strings.ContainsAny(text, "\r\n\t") {
-		return true
-	}
-	if len([]rune(strings.TrimSpace(text))) >= 24 {
-		return true
-	}
-	return false
-}
