+++ src/agent/internal/agent/tools_hid.go
@@ -787,14 +787,11 @@ func (t *KeyboardTapTool) Call(ctx context.Context, input string) (string, error
 		}
 	}
 
-	textInputLogf("keyboard_tap HID begin requested=%v resolved=%v modifier=0x%02x keycodes=%v hold_ms=%d pointer_mode=%s isolation_requested=%t", args.Keys, resolved.Keys, modifier, keys, holdMs, t.pointerModeOrDefault(), modifier != 0)
 	if err := t.iosKeyboardIsolation.withKeyboard(ctx, modifier != 0, func() error {
 		return t.tapKeyboardChord(modifier, keys, holdMs)
 	}); err != nil {
-		textInputLogf("keyboard_tap HID failed requested=%v modifier=0x%02x keycodes=%v err=%v", args.Keys, modifier, keys, err)
 		return toolErrorResultf(ctx, CodeToolExecutionFailed, "%v", err), nil
 	}
-	textInputLogf("keyboard_tap HID complete requested=%v modifier=0x%02x keycodes=%v", args.Keys, modifier, keys)
 	return "ok", nil
 }
 
@@ -908,10 +905,10 @@ func (t *KeyboardTextTool) Description() string {
 	return `US-keyboard ASCII text input only via USB HID physical keyboard (not the on-screen soft keyboard). ` +
 		`Allowed characters: a-z, A-Z, 0-9, space, and common US-keyboard punctuation. ` +
 		`For model/tool calls, pass JSON only, for example {"text":"App Store"}; do not pass a bare string. ` +
-		`Do NOT pass non-ASCII text, emoji, or spaced romanization — use enter_text_in_field for input box entry. ` +
-		`Do not transliterate Chinese/CJK targets to pinyin or guessed ASCII keywords; if enter_text_in_field is unavailable, report the blocker instead. ` +
+		`Do NOT pass non-ASCII text, emoji, or spaced romanization — use enter_text for input box entry. ` +
+		`Do not transliterate Chinese/CJK targets to pinyin or guessed ASCII keywords; if enter_text is unavailable, report the blocker instead. ` +
 		`Do not use keyboard_text for picker/wheel values, even if tapping the selected row appears to expose edit mode; use wheel_nudge and verify each returned screenshot instead. ` +
-		`keyboard_text remains for simple standalone ASCII typing outside the enter_text_in_field workflow. ` +
+		`keyboard_text remains for simple standalone ASCII typing outside the enter_text workflow. ` +
 		`Bare plain text is accepted only as a legacy compatibility fallback.`
 }
 
@@ -941,12 +938,12 @@ func (t *KeyboardTextTool) Call(ctx context.Context, input string) (string, erro
 		return toolErrorResultf(
 			ctx,
 			CodeInvalidArguments,
-			"keyboard_text supports only US-keyboard ASCII characters; unsupported characters: %q. Use enter_text_in_field for this target.",
+			"keyboard_text supports only US-keyboard ASCII characters; unsupported characters: %q. Use enter_text for this target.",
 			string(unsupported),
 		), nil
 	}
 	if looksLikeSpacedRomanizationBlob(text) {
-		return toolErrorResultString(ctx, CodeInvalidArguments, "keyboard_text received spaced romanization; use enter_text_in_field instead."), nil
+		return toolErrorResultString(ctx, CodeInvalidArguments, "keyboard_text received spaced romanization; use enter_text instead."), nil
 	}
 
 	if t.adb != nil {
